Module Compare.List_lengths

val (=) : 'a list -> 'b list -> bool

Compare.List_lengths.(xs = ys) iff xs and ys have the same length. In other words, iff Stdlib.List.compare_lengths xs ys = 0. Note that, like compare_lengths, this comparison only explores the lists up to the length of the shortest one.

val (<>) : 'a list -> 'b list -> bool

Compare.List_lengths.(xs <> ys) iff xs and ys have different lengths. In other words, iff Stdlib.List.compare_lengths xs ys <> 0. Note that, like compare_lengths, this comparison only explores the lists up to the length of the shortest one.

val (<) : 'a list -> 'b list -> bool

Compare.List_lengths.(xs < ys) iff xs is strictly shorter than ys. In other words, iff Stdlib.List.compare_lengths xs ys < 0. Note that, like compare_lengths, this comparison only explores the lists up to the length of the shortest one.

val (<=) : 'a list -> 'b list -> bool

Compare.List_lengths.(xs <= ys) iff xs is shorter than ys. In other words, iff Stdlib.List.compare_lengths xs ys <= 0. Note that, like compare_lengths, this comparison only explores the lists up to the length of the shortest one.

val (>=) : 'a list -> 'b list -> bool

Compare.List_lengths.(xs >= ys) iff xs is longer than ys. In other words, iff Stdlib.List.compare_lengths xs ys >= 0. Note that, like compare_lengths, this comparison only explores the lists up to the length of the shortest one.

val (>) : 'a list -> 'b list -> bool

Compare.List_lengths.(xs > ys) iff xs is strictly longer than ys. In other words, iff Stdlib.List.compare_lengths xs ys > 0. Note that, like compare_lengths, this comparison only explores the lists up to the length of the shortest one.

val compare : 'a list -> 'b list -> int

Compare.List_lengths.compare is an alias for Stdlib.List.compare_lengths.

val equal : 'a list -> 'b list -> bool

Compare.List_lengths.equal is an alias for Compare.List_lengths.( = ).