Module Compare.List_length_with

val (=) : 'a list -> int -> bool

Compare.List_length_with.(l = n) iff l is of length n. In other words iff Stdlib.List.compare_length_with l n = 0. Note that, like compare_length_with, this comparison does not explore the list l beyond its n-th element.

val (<>) : 'a list -> int -> bool

Compare.List_length_with.(l <> n) iff l is not of length n. In other words iff Stdlib.List.compare_length_with l n <> 0. Note that, like compare_length_with, this comparison does not explore the list l beyond its n-th element.

val (<) : 'a list -> int -> bool

Compare.List_length_with.(l < n) iff l is of length strictly less than n. In other words iff Stdlib.List.compare_length_with l n < 0. Note that, like compare_length_with, this comparison does not explore the list l beyond its n-th element.

val (<=) : 'a list -> int -> bool

Compare.List_length_with.(l <= n) iff l is of length less than n. In other words iff Stdlib.List.compare_length_with l n <= 0. Note that, like compare_length_with, this comparison does not explore the list l beyond its n-th element.

val (>=) : 'a list -> int -> bool

Compare.List_length_with.(l >= n) iff l is of length greater than n. In other words iff Stdlib.List.compare_length_with l n >= 0. Note that, like compare_length_with, this comparison does not explore the list l beyond its n-th element.

val (>) : 'a list -> int -> bool

Compare.List_length_with.(l > n) iff l is of length strictly greater than n. In other words iff Stdlib.List.compare_length_with l n > 0. Note that, like compare_length_with, this comparison does not explore the list l beyond its n-th element.

val compare : 'a list -> int -> int

Compare.List_length_with.compare is an alias for Stdlib.List.compare_length_with.

val equal : 'a list -> int -> bool

Compare.List_length_with.equal is an alias for Compare.List_length_with.( = ).