Module Env.Context

View over the context store, restricted to types, access and functional manipulation of an existing context.

type depth = [

| `Eq of int

| `Le of int

| `Lt of int

| `Ge of int

| `Gt of int

]

The tree depth of a fold. See the fold function for more information.

module type VIEW = sig ... end

module Kind : sig ... end

module type TREE = sig ... end

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes.

include VIEW with type key = string list and type value = bytes

type t

The type for context views.

type key = string list

The type for context keys.

type value = bytes

The type for context values.

type tree

The type for context trees.

Getters

val mem : t -> key -> bool Lwt.t

mem t k is an Lwt promise that resolves to true iff k is bound to a value in t.

val mem_tree : t -> key -> bool Lwt.t

mem_tree t k is like mem but for trees.

val find : t -> key -> value option Lwt.t

find t k is an Lwt promise that resolves to Some v if k is bound to the value v in t and None otherwise.

val find_tree : t -> key -> tree option Lwt.t

find_tree t k is like find but for trees.

val list : t -> ?offset:int -> ?length:int -> key -> (string * tree) list Lwt.t

list t key is the list of files and sub-nodes stored under k in t. The result order is not specified but is stable.

offset and length are used for pagination.

Setters

val add : t -> key -> value -> t Lwt.t

add t k v is an Lwt promise that resolves to c such that:

k is bound to v in c;

and c is similar to t otherwise.

If k was already bound in t to a value that is physically equal to v, the result of the function is a promise that resolves to t. Otherwise, the previous binding of k in t disappears.

val add_tree : t -> key -> tree -> t Lwt.t

add_tree is like add but for trees.

val remove : t -> key -> t Lwt.t

remove t k v is an Lwt promise that resolves to c such that:

k is unbound in c;

and c is similar to t otherwise.

Folding

val fold : ?depth:depth -> t -> key -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(key -> tree -> 'a -> 'a Lwt.t) -> 'a Lwt.t

fold ?depth t root ~init ~f recursively folds over the trees and values of t. The f callbacks are called with a key relative to root. f is never called with an empty key for values; i.e., folding over a value is a no-op.

Elements are traversed in lexical order of keys.

The depth is 0-indexed. If depth is set (by default it is not), then f is only called when the conditions described by the parameter is true:

Eq d folds over nodes and contents of depth exactly d.

Lt d folds over nodes and contents of depth strictly less than d.

Le d folds over nodes and contents of depth less than or equal to d.

Gt d folds over nodes and contents of depth strictly more than d.

Ge d folds over nodes and contents of depth more than or equal to d.

If order is `Sorted (the default), the elements are traversed in lexicographic order of their keys. For large nodes, these two modes are memory-consuming, use `Undefined for a more memory efficient fold.

module Tree : TREE with type t := t and type key := key and type value := value and type tree := tree

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes.

val register_resolver : 'a Base58.encoding -> (t -> string -> 'a list Lwt.t) -> unit

val complete : t -> string -> string list Lwt.t

val get_hash_version : t -> Context_hash.Version.t

Get the hash version used for the context

val set_hash_version : t -> Context_hash.Version.t -> t Error_monad.shell_tzresult Lwt.t

Set the hash version used for the context. It may recalculate the hashes of the whole context, which can be a long process. Returns an Error if the hash version is unsupported.

type cache_key

type cache_value = ..

module type CACHE = sig ... end

module Cache : CACHE with type t := t and type size := int and type index := int and type identifier := string and type key = cache_key and type value = cache_value