Module Env.RPC_answer

type 'o t = [

| `Ok of 'o

| `OkChunk of 'o

| `OkStream of 'o stream

| `Created of string option

| `No_content

| `Unauthorized of Error_monad.error list option

| `Forbidden of Error_monad.error list option

| `Not_found of Error_monad.error list option

| `Conflict of Error_monad.error list option

| `Error of Error_monad.error list option

]

Return type for service handler

and 'a stream = {

next : unit -> 'a option Lwt.t;

shutdown : unit -> unit;

}

val return : 'o -> 'o t Lwt.t

val return_chunked : 'o -> 'o t Lwt.t

return_chunked is identical to return but it indicates to the server that the result might be long and that the serialisation should be done in mutliple chunks.

You should use return_chunked when returning an (unbounded or potentially large) list, array, map, or other such set.

val return_stream : 'o stream -> 'o t Lwt.t

val not_found : 'o t Lwt.t

val fail : Error_monad.error list -> 'a t Lwt.t