Parameter PROTOCOL_V8.Env

module CamlinternalFormatBasics : module type of struct include Tezos_protocol_environment_sigs_internals.CamlinternalFormatBasics end

module Pervasives : sig ... end

The OCaml Standard library.

module Either : sig ... end

Either type.

module String : sig ... end

String operations.

module Char : sig ... end

Character operations.

module Bytes : sig ... end

Byte sequence operations.

module Int32 : sig ... end

32-bit integers.

module Int64 : sig ... end

64-bit integers.

module Format : sig ... end

Pretty-printing.

module Logging : sig ... end

module Hex : sig ... end

Hexadecimal encoding.

module Z : sig ... end

Integers.

module Q : sig ... end

Rationals.

module Lwt : sig ... end

module Data_encoding : sig ... end

module Raw_hashes : sig ... end

module Compare : sig ... end

module Time : sig ... end

module TzEndian : sig ... end

module Bits : sig ... end

module Equality_witness : sig ... end

This module provides support for type equalities and runtime type identifiers.

module FallbackArray : sig ... end

This module implements arrays equipped with accessors that cannot raise exceptions. Reading out of the bounds of the arrays return a fallback value fixed at array construction time, writing out of the bounds of the arrays is ignored.

module Error_monad : sig ... end

module Seq : sig ... end

module List : sig ... end

module Array : sig ... end

This module is a very restricted subset of OCaml's Stdlib Array module. There is just enough exposed that you can pass arrays around to some functions and such, but not enough that you can easily shoot yourself in the foot. (See details below.)

module Set : sig ... end

module Map : sig ... end

module Option : sig ... end

Signature from Lwtreslib's option module

module Result : sig ... end

module RPC_arg : sig ... end

See src/lib_rpc/RPC_arg.mli for documentation

module RPC_path : sig ... end

module RPC_query : sig ... end

module RPC_service : sig ... end

module RPC_answer : sig ... end

module RPC_directory : sig ... end

module Base58 : sig ... end

module S : sig ... end

module Blake2B : sig ... end

Builds a new Hash type using Blake2B.

module Bls : sig ... end

Tezos - BLS12-381 cryptography

module Ed25519 : sig ... end

Tezos - Ed25519 cryptography

module Secp256k1 : sig ... end

Tezos - Secp256k1 cryptography

module P256 : sig ... end

Tezos - P256 cryptography

module Chain_id : sig ... end

module Signature : sig ... end

module Block_hash : sig ... end

module Operation_hash : sig ... end

module Operation_list_hash : sig ... end

module Operation_list_list_hash : sig ... end

module Protocol_hash : sig ... end

module Context_hash : sig ... end

module Sapling : sig ... end

module Timelock : sig ... end

module Vdf : sig ... end

module Micheline : sig ... end

module Block_header : sig ... end

module Bounded : sig ... end

This module implements bounded (or refined) versions of data types.

module Fitness : sig ... end

module Operation : sig ... end

Tezos operations.

module Context : sig ... end

View over the context store, restricted to types, access and functional manipulation of an existing context.

module Updater : sig ... end

Tezos Protocol Environment - Protocol updater.

module RPC_context : sig ... end

module Wasm_2_0_0 : sig ... end

module Plonk : sig ... end

aPlonK is a PlonK-based proving system. As such, it provides a way to create succinct cryptographic proofs about a given predicate, which can be then verified with a low computational cost.

module Dal : sig ... end